mynotes_project/requirements.txt
django==5.2.1
psycopg2-binary
markdown
django-crispy-forms
crispy-bootstrap5